Completed part 4
