+++ .replit
@@ -1,2 +1,2 @@
 language = "cpp"
-run = "clang++ -std=c++17 -Weverything -Wno-c++98-compat -Wno-padded -Wno-deprecated -Wno-missing-prototypes -Wno-exit-time-destructors main.cpp && ./a.out"
+run = "clang++ -std=c++17 -Weverything -Wno-c++98-compat -Wno-padded -Wno-deprecated -Wno-missing-prototypes -Wno-exit-time-destructors main.cpp Camera.cpp Kitchen.cpp Laptop.cpp OvenStove.cpp System.cpp Wrappers.cpp && ./a.out"
